Refactor button components
Move Button and CopyButton under a generic atom.
Abstract Button into RegularButton internally.
Move stories under one heading.
Update designs to work for Button.
