+++ src/atoms/_index.scss
@@ -1,5 +1,4 @@
 @import './button';
-@import './copyButton';
 @import './expertise';
 @import './search';
 @import './tag';
