+++ src/atoms/button/copyButton/_index.scss
