+++ src/atoms/button/copyButton/docs.mdx
@@ -8,7 +8,7 @@ import {
   ColorItem,
   Meta,
 } from '@storybook/addon-docs/blocks';
-import CopyButton from 'atoms/copyButton';
+import { CopyButton } from 'atoms/button';
 
 # CopyButton
