+++ src/atoms/button/regularButton/_index.scss
@@ -1,0 +1,45 @@
+// SCSS variables - used to color the icon
+$button-fore-color: #FFFFFF;
+
+// Colors
+:root {
+  --button-back-color: #3B3EFC;
+  --button-fore-color: #{$button-fore-color};
+  --button-shadow-color: rgba(0, 0, 0, 0.25);
+  --button-shadow-focus-color: rgba(0, 0, 0, 0.29);
+  --button-back-color-secondary: #DC325F;
+}
+
+.btn, a.btn {
+  display: inline-block;
+  padding: 0.625rem 0.875rem;
+  margin-top: 0.75rem;
+  background: var(--button-back-color);
+  color: var(--button-fore-color);
+  cursor: pointer;
+  border: none;
+  border-radius: 1rem;
+  font-weight: 500;
+  font-size: 1.125rem;
+  line-height: 1.4;
+  transition: 0.3s ease all;
+  box-shadow:  0px 4px 8px var(--button-shadow-color);
+  &:hover, &:focus {
+    box-shadow: 0px 8px 16px var(--button-shadow-color);
+    text-decoration: none;
+    outline: none;
+  }
+
+  svg {
+    vertical-align: sub;
+  }
+
+  @media screen and (min-width: $layout-large-breakpoint) {
+    font-size: 1.375rem;
+    line-height: 1.35;
+  }
+
+  &.secondary {
+    background: var(--button-back-color-secondary);
+  }
+}
