+++ src/atoms/button/regularButton/docs.mdx
@@ -8,7 +8,7 @@ import {
   ColorItem,
   Meta,
 } from '@storybook/addon-docs/blocks';
-import Button from 'atoms/button';
+import { Button } from 'atoms/button';
 
 # Button
